Fix arguments and output of text format
PHP_CodeCoverage_Report_Text expects different args for both constructor and process method. It also never saves to a file, but it returns the report string instead.

To avoid having to do addititonal processing in the Listener, in text case simply output the report. It is very useful to be able to get a quick report this way directly on the screen.